requirements.txt
Django==1.7.5
MySQL-python==1.2.5
django-stronghold==0.2.7
wsgiref==0.1.2
